requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
Naked==0.1.31
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
urllib3==1.26.6
